Refatorada a estrutura da pasta components para conter apenas os widgets filhos diretos da pagina
